// src/interfaces/OrderItem.ts
export default interface OrderItem {
    itemCode: string;
    priceLevel: string;
    price: string;
    brand: string;
    pack: string;
    size: string;
    uom: string;
    description: string;
    upc: string;
    unit: string;
    deal: string;
    dealOff: string;
    status: string;
    date: string;
}